dispenser/requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dlib==19.20.0
face-recognition==1.3.0
face-recognition-models==0.3.0
idna==2.10
numpy==1.19.1
opencv-python==4.3.0.36
otpauth==1.0.1
Pillow==7.2.0
pyserial==3.4
requests==2.24.0
schedule==0.6.0
urllib3==1.25.10